requirements.txt
pyspark==3.5.0
psycopg2-binary==2.9.7
pandas==2.1.0
numpy==1.27.0
sqlalchemy==2.0.20
apache-airflow==2.8.0
kafka-python==2.1.0
elasticsearch==8.12.0
matplotlib==3.8.0
seaborn==0.13.2
PyYAML==6.0
